moved calculate_enhancement to dnpTools
-moved calculate_enhancement to dnpTools
-minor fix in dnpNMR.integrate
-minor axis label change in GUI
-added tests
