+++ dnplab/__init__.py
@@ -2,8 +2,8 @@
 from .dnpData import dnpdata, dnpdata_collection, create_workspace
 from .dnpImport import load
 from .dnpSave import save
-from . import dnpNMR
 from . import dnpIO
+from . import dnpNMR
 from . import dnpFit
 from . import dnpTools
 from . import dnpHydration
