+++ dnplab/dnpNMR.py
@@ -297,19 +297,17 @@ def integrate(all_data, dim="t2", integrate_center=0, integrate_width="full"):
     data, isDict = return_data(all_data)
 
     if integrate_width == "full":
-        integrateMin = min(data.coords[dim])
-        integrateMax = max(data.coords[dim])
+        pass
     elif isinstance(integrate_width, int) or isinstance(integrate_width, float):
         integrateMin = integrate_center - _np.abs(integrate_width) / 2.0
         integrateMax = integrate_center + _np.abs(integrate_width) / 2.0
+        data = data[dim, (integrateMin, integrateMax)]
 
     proc_parameters = {
         "dim": dim,
         "integrate_center": integrate_center,
         "integrate_width": integrate_width,
     }
-
-    data = data[dim, (integrateMin, integrateMax)]
 
     index = data.index(dim)
     data.values = _np.trapz(data.values, x=data.coords[dim], axis=index)
